requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
greenlet==2.0.2
idna==3.4
psycopg2-binary==2.9.5
pyTelegramBotAPI==4.10.0
python-dotenv==0.21.1
redis==4.5.1
requests==2.28.2
SQLAlchemy==2.0.4
SQLAlchemy-Utils==0.40.0
typing_extensions==4.5.0
urllib3==1.26.14
